requirements.txt
requests
google-cloud-pubsub
google-cloud-bigquery
